Timer Update
Added to the class Timer(PhaseThread) to make some changes so that we can set the timer values. Worked with formatting, Loop order, initialization, and explosion. Includes the def _update(self), def __str__(self). Recognized that def run(self) is the the entry point for the timer when it starts.