Keypad Adjustments
Updated the Keypad class so that it works better for when the users have to type in a number pattern. Now, it only accepts a limited number of digits, which helps prevent mistakes or too much input. If the user enters the wrong code or gets it right, the keypad will stop taking input so the game can respond. I also added options to pause the keypad if needed and to check what the user has typed so far, which can be useful for showing the input on the screen.