Start Game Graphic
I added a new start screen that shows up when the game launches. It has the game title and a "Continue" button that players can click to begin. This makes the game feel more polished and gives players a clear place to start. I also connected it to the main game window so everything flows smoothly. Class added to bomb_phases.py and def start_main_game(): added to bomb.py.