Format Changes with Tk. Panel Changes Improves
In this update, all classes were converted to tk format to ensure full compatibility and visibility within the GUI system. Additionally, a major change was made to bomb.py, where it now imports from the phases module using a different structure. This adjustment made it possible to integrate and control the bomb defusal process through a physical button on the Raspberry Pi panel, allowing for a smoother connection between the software and hardware components.