ADDING GATE 4 IMAGE AND FUNCTIONS
gate 4 is a triangle puzzle to find out how many figures are being shown. the player will use the keypad to input their answer. i input this into the code and added the image for the puzzle and the functions for the game to read the players input and correct answer.